+++ CHANGELOG.md
@@ -26,6 +26,15 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/infrastructure/branches/compare/master%0Drelease-4.7.0
 
+### Added
+- Support for Jira Service Desk installer. Unlocks JPERF-277.
+
+## Deprecated
+- Deprecate `ProductDistribution` in favor of `ProductDistribution` in `distribution` package.
+- Deprecate `PublicJiraSoftwareDistributions` in favor of `PublicJiraSoftwareDistribution`.
+
+[JPERF-277]: https://ecosystem.atlassian.net/browse/JPERF-277
+
 ## [4.7.0] - 2019-01-29
 [4.7.0]: https://bitbucket.org/atlassian/infrastructure/branches/compare/release-4.7.0%0Drelease-4.6.1